requirements.txt
pillow
requests
font-hanken-grotesk
schedule
inky[rpi]
jpholiday
